backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
garminconnect==0.2.22
openai==1.30.1
sqlalchemy==2.0.30
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
python-multipart==0.0.9
